requirements.txt
openai==1.14.3
chromadb==0.4.21
sentence-transformers==2.2.2
transformers==4.36.2
pandas==2.2.1
pyarrow==15.0.2
streamlit==1.33.0
python-dotenv==1.0.1
